refactor: further decompose TrailRunningWorkout into smaller components
- Extract SectionTypeSelector component (200+ lines)
- Extract RepeatIntervalsForm component (200+ lines)
- Extract SectionsList component (300+ lines)
- Extract SectionForm component (50+ lines)
- Extract SectionFormCard component (200+ lines)
- Extract smartDefaults utility (100+ lines)
- Reduce TrailRunningWorkout from 1,255 to 126 lines (90% reduction)
- Maintain all functionality while improving maintainability
- Fix TypeScript and linting issues
